requirements.txt
# Ekstraksi dokumen
python-docx==1.1.2
pdfplumber==0.11.4
PyMuPDF==1.24.9
Pillow==10.4.0

# OCR
pytesseract==0.3.13

# AI APIs
anthropic==0.34.2
google-generativeai==0.7.2

# Google Workspace
gspread==6.1.2
google-auth==2.34.0
google-api-python-client==2.143.0

# Utilitas
python-dotenv==1.0.1
requests==2.32.3
